ops-fault-diagnosis/requirements.txt
anthropic>=0.97.0
rich>=14.0
pydantic>=2.0
click>=8.0
python-dotenv>=1.0
